# pytorrent/openapi/openapi.json
{
  "components": {
    "schemas": {
      "ApiError": {
        "properties": {
          "error": {
            "type": "string"
          },
          "ok": {
            "enum": [
              false
            ],
            "type": "boolean"
          }
        },
        "required": [
          "ok",
          "error"
        ],
        "type": "object"
      },
      "ApiOk": {
        "properties": {
          "ok": {
            "enum": [
              true
            ],
            "type": "boolean"
          }
        },
        "required": [
          "ok"
        ],
        "type": "object"
      },
...
      "DeletedResponse": {
        "allOf": [
          {
            "$ref": "#/components/schemas/ApiOk"
          },
          {
            "properties": {
              "deleted": {
                "type": "integer"
              }
            },
            "type": "object"
          }
        ]
      },
...
      "OperationLogClearRequest": {
        "properties": {
          "event_type": {
            "description": "Optional event type filter. Empty clears all active-profile operation logs.",
            "type": "string"
          },
          "profile_id": {
            "type": "integer",
            "default": 1,
            "minimum": 1,
            "description": "Optional rTorrent profile id. Overrides active profile for this request."
          },
          "profile_name": {
            "type": "string",
            "description": "Optional rTorrent profile name. Used only when profile_id is not provided."
          }
        },
        "type": "object"
      },
      "OperationLogEntry": {
        "properties": {
          "action": {
            "nullable": true,
            "type": "string"
          },
          "created_at": {
            "type": "string"
          },
          "details": {
            "additionalProperties": true,
            "type": "object"
          },
          "details_h": {
            "type": "string"
          },
          "details_json": {
            "type": "string"
          },
          "event_type": {
            "type": "string"
          },
          "id": {
            "type": "integer"
          },
          "message": {
            "type": "string"
          },
          "profile_id": {
            "nullable": true,
            "type": "integer"
          },
          "severity": {
            "type": "string"
          },
          "source": {
            "type": "string"
          },
          "torrent_hash": {
            "nullable": true,
            "type": "string"
          },
          "torrent_name": {
            "nullable": true,
            "type": "string"
          },
          "user_id": {
            "type": "integer"
          }
        },
        "type": "object"
      },
      "OperationLogRetentionResult": {
        "properties": {
          "deleted": {
            "type": "integer"
          },
          "deleted_days": {
            "type": "integer"
          },
          "deleted_lines": {
            "type": "integer"
          },
          "settings": {
            "$ref": "#/components/schemas/OperationLogSettings"
          }
        },
        "type": "object"
      },
      "OperationLogSettings": {
        "properties": {
          "created_at": {
            "type": "string"
          },
          "profile_id": {
            "type": "integer"
          },
          "retention_days": {
            "maximum": 3650,
            "minimum": 1,
            "type": "integer"
          },
          "retention_lines": {
            "maximum": 1000000,
            "minimum": 100,
            "type": "integer"
          },
          "retention_mode": {
            "enum": [
              "days",
              "lines",
              "both",
              "manual"
            ],
            "type": "string"
          },
          "updated_at": {
            "type": "string"
          },
          "user_id": {
            "type": "integer"
          }
        },
        "type": "object"
      },
      "OperationLogSettingsRequest": {
        "properties": {
          "retention_days": {
            "maximum": 3650,
            "minimum": 1,
            "type": "integer"
          },
          "retention_lines": {
            "maximum": 1000000,
            "minimum": 100,
            "type": "integer"
          },
          "retention_mode": {
            "enum": [
              "days",
              "lines",
              "both",
              "manual"
            ],
            "type": "string"
          },
          "profile_id": {
            "type": "integer",
            "default": 1,
            "minimum": 1,
            "description": "Optional rTorrent profile id. Overrides active profile for this request."
          },
          "profile_name": {
            "type": "string",
            "description": "Optional rTorrent profile name. Used only when profile_id is not provided."
          }
        },
        "type": "object"
      },
      "OperationLogSettingsResponse": {
        "allOf": [
          {
            "$ref": "#/components/schemas/ApiOk"
          },
          {
            "properties": {
              "retention": {
                "$ref": "#/components/schemas/OperationLogRetentionResult"
              },
              "settings": {
                "$ref": "#/components/schemas/OperationLogSettings"
              }
            },
            "type": "object"
          }
        ]
      },
      "OperationLogStats": {
        "properties": {
          "by_day": {
            "items": {
              "properties": {
                "bucket": {
                  "type": "string"
                },
                "n": {
                  "type": "integer"
                }
              },
              "type": "object"
            },
            "type": "array"
          },
          "by_month": {
            "items": {
              "properties": {
                "bucket": {
                  "type": "string"
                },
                "n": {
                  "type": "integer"
                }
              },
              "type": "object"
            },
            "type": "array"
          },
          "by_type": {
            "items": {
              "properties": {
                "event_type": {
                  "type": "string"
                },
                "n": {
                  "type": "integer"
                }
              },
              "type": "object"
            },
            "type": "array"
          },
          "settings": {
            "$ref": "#/components/schemas/OperationLogSettings"
          },
          "top_actions": {
            "items": {
              "properties": {
                "action": {
                  "type": "string"
                },
                "n": {
                  "type": "integer"
                }
              },
              "type": "object"
            },
            "type": "array"
          },
          "total": {
            "type": "integer"
          }
        },
        "type": "object"
      },
      "OperationLogsResponse": {
        "allOf": [
          {
            "$ref": "#/components/schemas/ApiOk"
          },
          {
            "properties": {
              "limit": {
                "type": "integer"
              },
              "logs": {
                "items": {
                  "$ref": "#/components/schemas/OperationLogEntry"
                },
                "type": "array"
              },
              "offset": {
                "type": "integer"
              },
              "settings": {
                "$ref": "#/components/schemas/OperationLogSettings"
              },
              "stats": {
                "$ref": "#/components/schemas/OperationLogStats"
              },
              "total": {
                "type": "integer"
              }
            },
            "type": "object"
          }
        ]
      },
...
      "PlannerHistoryEntry": {
        "additionalProperties": true,
        "properties": {
          "at": {
            "format": "date-time",
            "type": "string"
          },
          "count": {
            "type": "integer"
          },
          "down": {
            "type": "integer"
          },
          "dry_run": {
            "type": "boolean"
          },
          "event": {
            "type": "string"
          },
          "pause_reason": {
            "type": "string"
          },
          "paused": {
            "type": "integer"
          },
          "reasons": {
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "resumed": {
            "type": "integer"
          },
          "up": {
            "type": "integer"
          }
        },
        "type": "object"
      },
      "PlannerPreviewResponse": {
        "allOf": [
          {
            "$ref": "#/components/schemas/ApiOk"
          },
          {
            "properties": {
              "history": {
                "items": {
                  "$ref": "#/components/schemas/PlannerHistoryEntry"
                },
                "type": "array"
              },
              "history_total": {
                "type": "integer"
              },
              "preview": {
                "additionalProperties": true,
                "type": "object"
              }
            },
            "type": "object"
          }
        ]
      },
...
    },
    "securitySchemes": {
      "apiKeyAuth": {
        "description": "Optional per-user API token. Tokens are accepted only when PYTORRENT_AUTH_ENABLE=1. When auth is disabled, API endpoints do not require credentials.",
        "in": "header",
        "name": "X-API-Key",
        "type": "apiKey"
      },
      "bearerApiToken": {
        "description": "Optional per-user API token in Authorization: Bearer <token>. Tokens are accepted only when PYTORRENT_AUTH_ENABLE=1.",
        "scheme": "bearer",
        "type": "http"
      },
      "sessionCookie": {
        "description": "Optional Flask session cookie. Required only when PYTORRENT_AUTH_ENABLE=1 and no API token is used.",
        "in": "cookie",
        "name": "session",
        "type": "apiKey"
      }
    },
    "parameters": {
      "ProfileId": {
        "name": "profile_id",
        "in": "query",
        "required": false,
        "description": "Optional rTorrent profile id for profile-scoped API operations. If omitted, the user active profile is used for backward compatibility; fallback profile id 1 is used only when no active profile exists and the user can access it.",
        "schema": {
          "type": "integer",
          "default": 1,
          "minimum": 1
        }
      },
      "ProfileName": {
        "name": "profile_name",
        "in": "query",
        "required": false,
        "description": "Optional rTorrent profile name for profile-scoped API operations. profile_id has priority when both are supplied.",
        "schema": {
          "type": "string"
        }
      }
    }
  },
  "info": {
    "title": "pyTorrent API",
    "version": "0.0.1"
  },
  "openapi": "3.0.3",
  "paths": {
...
    "/api/download-planner": {
      "get": {
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            },
            "description": "OK"
          }
        },
        "summary": "Manage download planner settings",
        "parameters": [
          {
            "$ref": "#/components/parameters/ProfileId"
          },
          {
            "$ref": "#/components/parameters/ProfileName"
          }
        ]
      },
      "post": {
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "profile_id": {
                    "type": "integer",
                    "default": 1,
                    "minimum": 1,
                    "description": "Optional rTorrent profile id. Overrides active profile for this request."
                  },
                  "profile_name": {
                    "type": "string",
                    "description": "Optional rTorrent profile name. Used only when profile_id is not provided."
                  }
                }
              }
            }
          },
          "required": false
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            },
            "description": "OK"
          }
        },
        "summary": "Manage download planner settings",
        "parameters": [
          {
            "$ref": "#/components/parameters/ProfileId"
          },
          {
            "$ref": "#/components/parameters/ProfileName"
          }
        ]
      }
    },
    "/api/download-planner/check": {
      "post": {
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "profile_id": {
                    "type": "integer",
                    "default": 1,
                    "minimum": 1,
                    "description": "Optional rTorrent profile id. Overrides active profile for this request."
                  },
                  "profile_name": {
                    "type": "string",
                    "description": "Optional rTorrent profile name. Used only when profile_id is not provided."
                  }
                }
              }
            }
          },
          "required": false
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            },
            "description": "OK"
          }
        },
        "summary": "Run download planner check",
        "parameters": [
          {
            "$ref": "#/components/parameters/ProfileId"
          },
          {
            "$ref": "#/components/parameters/ProfileName"
          }
        ]
      }
    },
    "/api/download-planner/history": {
      "delete": {
        "description": "Deletes Download Planner action history for the active profile. Saved Planner settings are preserved.",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiOk"
                    },
                    {
                      "properties": {
                        "deleted": {
                          "type": "integer"
                        },
                        "history": {
                          "items": {
                            "$ref": "#/components/schemas/PlannerHistoryEntry"
                          },
                          "type": "array"
                        },
                        "history_total": {
                          "type": "integer"
                        }
                      },
                      "type": "object"
                    }
                  ]
                }
              }
            },
            "description": "History cleared"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "No active profile"
          }
        },
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "summary": "Clear Planner action history",
        "parameters": [
          {
            "$ref": "#/components/parameters/ProfileId"
          },
          {
            "$ref": "#/components/parameters/ProfileName"
          }
        ]
      }
    },
    "/api/download-planner/override": {
      "post": {
        "description": "Set a temporary manual planner override.",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "down_limit": {
                    "type": "integer"
                  },
                  "enabled": {
                    "type": "boolean"
                  },
                  "minutes": {
                    "type": "integer"
                  },
                  "up_limit": {
                    "type": "integer"
                  },
                  "profile_id": {
                    "type": "integer",
                    "default": 1,
                    "minimum": 1,
                    "description": "Optional rTorrent profile id. Overrides active profile for this request."
                  },
                  "profile_name": {
                    "type": "string",
                    "description": "Optional rTorrent profile name. Used only when profile_id is not provided."
                  }
                },
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "ok": {
                      "type": "boolean"
                    }
                  },
                  "type": "object"
                }
              }
            },
            "description": "OK"
          }
        },
        "summary": "Set download planner override",
        "parameters": [
          {
            "$ref": "#/components/parameters/ProfileId"
          },
          {
            "$ref": "#/components/parameters/ProfileName"
          }
        ]
      }
    },
    "/api/download-planner/preview": {
      "get": {
        "description": "Preview matching planner rule and next action.",
        "parameters": [
          {
            "description": "Number of Planner history rows to return.",
            "in": "query",
            "name": "history_limit",
            "required": false,
            "schema": {
              "maximum": 200,
              "minimum": 1,
              "type": "integer"
            }
          },
          {
            "$ref": "#/components/parameters/ProfileId"
          },
          {
            "$ref": "#/components/parameters/ProfileName"
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlannerPreviewResponse"
                }
              }
            },
            "description": "Planner preview"
          }
        },
        "summary": "Preview download planner"
      }
    },
...
    "/api/operation-logs": {
      "get": {
        "description": "Lists active-profile operation logs with filters, retention settings and statistics.",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "maximum": 1000,
              "minimum": 1,
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "minimum": 0,
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "type",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "$ref": "#/components/parameters/ProfileId"
          },
          {
            "$ref": "#/components/parameters/ProfileName"
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OperationLogsResponse"
                }
              }
            },
            "description": "Operation logs"
          }
        },
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "summary": "List operation logs"
      }
    },
    "/api/operation-logs/apply-retention": {
      "post": {
        "description": "Runs retention cleanup for active-profile operation logs.",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiOk"
                    },
                    {
                      "$ref": "#/components/schemas/OperationLogRetentionResult"
                    }
                  ]
                }
              }
            },
            "description": "Retention result"
          }
        },
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "summary": "Apply operation log retention",
        "parameters": [
          {
            "$ref": "#/components/parameters/ProfileId"
          },
          {
            "$ref": "#/components/parameters/ProfileName"
          }
        ]
      }
    },
    "/api/operation-logs/clear": {
      "post": {
        "description": "Clears active-profile operation logs, optionally limited to one event type.",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OperationLogClearRequest"
              }
            }
          },
          "required": false
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeletedResponse"
                }
              }
            },
            "description": "Deleted count"
          }
        },
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "summary": "Clear operation logs",
        "parameters": [
          {
            "$ref": "#/components/parameters/ProfileId"
          },
          {
            "$ref": "#/components/parameters/ProfileName"
          }
        ]
      }
    },
    "/api/operation-logs/settings": {
      "post": {
        "description": "Saves active-profile operation log retention and applies it immediately.",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OperationLogSettingsRequest"
              }
            }
          },
          "required": false
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OperationLogSettingsResponse"
                }
              }
            },
            "description": "Saved settings"
          }
        },
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "summary": "Save operation log retention settings",
        "parameters": [
          {
            "$ref": "#/components/parameters/ProfileId"
          },
          {
            "$ref": "#/components/parameters/ProfileName"
          }
        ]
      }
    },
...
    "/api/poller/settings": {
      "get": {
        "description": "Get poller settings",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "ok": {
                      "type": "boolean"
                    }
                  },
                  "type": "object"
                }
              }
            },
            "description": "Get poller settings"
          }
        },
        "summary": "Get poller settings",
        "parameters": [
          {
            "$ref": "#/components/parameters/ProfileId"
          },
          {
            "$ref": "#/components/parameters/ProfileName"
          }
        ]
      },
      "post": {
        "description": "Save poller settings",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "ok": {
                      "type": "boolean"
                    }
                  },
                  "type": "object"
                }
              }
            },
            "description": "Save poller settings"
          }
        },
        "summary": "Save poller settings",
        "parameters": [
          {
            "$ref": "#/components/parameters/ProfileId"
          },
          {
            "$ref": "#/components/parameters/ProfileName"
          }
        ]
      }
    },
...
  }
}